requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
python-dotenv==1.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
pymongo==4.6.3

# Dependências de teste
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
pytest-html==3.2.0
pytest-mock==3.11.1
coverage==7.3.2
factory-boy==3.3.0
faker==19.6.2